+++ src/style.css
@@ -27,6 +27,7 @@ li:nth-child(odd) {
   letter-spacing: 8px;
   padding-bottom: 50px;
   font-weight: 900;
+  font-size: 50px;
   font-family: 'Faster One', cursive;
 }
 
@@ -52,13 +53,12 @@ li:nth-child(odd) {
 }
 
 .recent {
-  font-family: Georgia, 'Times New Roman', Times, serif;
+  font-family: 'Faster One', cursive;
   font-size: 20px;
   padding: 8px;
 }
 
 .refresh-btn {
-  padding: 10px;
   box-shadow: 1px 1px 5px 4px;
 }
 
@@ -77,7 +77,7 @@ li:nth-child(odd) {
 }
 
 .add-title {
-  font-family: sans-serif;
+  font-family: 'Faster One', cursive;
   padding-bottom: 10px;
 }
